pyproject.toml
[project]
name = "git-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django>=5.1.3",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "django-pytest>=0.2.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "ruff>=0.7.4",
]


[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true
skip_glob = ["tests/data", "profiling"]
known-third-party = [
    "django",
    "django-cors-headers",
    "djangorestframework",
    "djangorestframework-simplejwt",
]

[tool.ruff]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "N",   # pep8-naming
    "T20", # flake8-print
]

ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
]

extend-exclude = [".pyenv", ".vscode", "/usr/**", ".venv/**"]
